First upload of the word guessing game
The game takes a word from the words.txt file and the player is given five chances to guess the word. After each guess, the letters used are given a different color based on whether they were not in the word, in the word and in the wronge place, or in the word in the correct place. The game also informs the player of how many guesses they have left and a collection of letters that were misplaced or outright wrong.