Add files via upload
Fix for the duplicate letter bug. Program now keeps a counter of how many times a letter appears in the word and removes from the counter for each letter that is in the correct spot. When both have been guessed correctly, the value reaches zero and it is removed from the list of misplaced letters.